nep-63-fix-refresh-issue: restructured the navigation flow and resloved the refresh screen issues
