+++ src/frame/app-layout.js
@@ -1,17 +1,10 @@
 import React from 'react'
 import { Container } from 'native-base'
 
-import BottomNavigation from '../home/components/bottom-navigation'
-
 class AppLayout extends React.PureComponent {
 	render() {
 		const { children } = this.props
-		return (
-			<Container>
-				{children}
-				<BottomNavigation />
-			</Container>
-		)
+		return <Container>{children}</Container>
 	}
 }
